% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_plot.R
\name{payment_plot}
\alias{payment_plot}
\title{Payment Plot}
\usage{
payment_plot(data, pay_type)
}
\arguments{
\item{data}{a dataframe}

\item{pay_type}{a string name for either the Average Medicare Payments, the Average Total Payments, or the Average Covered Charges}
}
\value{
A boxplot of payments of \code{pay_type} by DRG code
}
\description{
This function creates a boxplot of payments for each \code{pay_type} by DRG code.
}
\examples{
payment_plot(data = sample_data, pay_type = 'Average Medicare Payments')

}
